% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauda-docs.R
\name{cauda.save}
\alias{cauda.save}
\title{Save a cauda plot as a high-resolution PNG}
\usage{
cauda.save(filename, expr, width = 3200, height = 2400, res = 220)
}
\arguments{
\item{filename}{Character. Output filename (e.g. "dag.png"). Saved to working directory.}

\item{expr}{The cauda plot expression to run (unquoted).}

\item{width}{Integer. Image width in pixels. Default is 3200.}

\item{height}{Integer. Image height in pixels. Default is 2400.}

\item{res}{Integer. Resolution in DPI. Default is 220.}
}
\value{
Invisibly returns the filename.
}
\description{
Reruns a cauda plot expression inside a PNG device to produce a
publication-quality image regardless of RStudio window size.
}
\examples{
\dontrun{
df_ready <- cauda.prep(mtcars, verbose = FALSE)
cauda.save("my_dag.png", cauda.dag(df_ready, highlight = "mpg", verbose = FALSE))
}

}
