% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CauDA.R
\name{caudabox_from_csvs}
\alias{caudabox_from_csvs}
\title{Generate a CauDAbox object from CSVs}
\usage{
caudabox_from_csvs(metadata, tableX, dictionary)
}
\arguments{
\item{metadata}{PLACEHOLDER}

\item{tableX}{PLACEHOLDER}

\item{dictionary}{PLACEHOLDER}
}
\description{
PLACEHOLDER
}
