% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mediation.R
\name{enframe_mediation}
\alias{enframe_mediation}
\title{Wrangle output from mediation analysis to table.}
\usage{
enframe_mediation(res_mediate)
}
\arguments{
\item{res_mediate}{The outcome from our mediation call}
}
\value{
A structured table summarizing the mediation call.
}
\description{
Make a table out of a mediate object.
}
