% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mediation.R
\name{fit_mediation}
\alias{fit_mediation}
\title{Run mediation analysis for a given feature pair.}
\usage{
fit_mediation(y, x_dir, x_indir, f, metadata)
}
\arguments{
\item{y}{The outcome variable}

\item{x_dir}{The putative direct effect variable .}

\item{x_indir}{The putative indirect effect variable.}

\item{f}{A formula object. Used to assess differential associations.}

\item{metadata}{Additional information such as group or covariate values.}
}
\value{
To be determined. Currently a table of mediation output summary
}
\description{
Should not be run on its own. Typically, the main workhorse function will handle this for you.
}
\examples{
mediation_df <- data.frame(z = sample(LETTERS[1:2], 32, replace = T), mtcars)
fit_mediation(y = rnorm(32), x_dir = rnorm(32), x_indir = rnorm(32), f = ~z, metadata = mediation_df)
  
}
