% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CauDA.R
\name{get_cols_to_include}
\alias{get_cols_to_include}
\title{Assesses which columns to consider for inclusion in regression}
\usage{
get_cols_to_include(caudabox, column)
}
\arguments{
\item{caudabox}{A caudabox object}

\item{column}{The feature of interest in tableX}
}
\value{
A vector of other features that affect the feature of interest
}
\description{
PLACEHOLDER
}
