% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CauDA.R
\name{RHS_from_edgemat}
\alias{RHS_from_edgemat}
\title{PLACEHOLDER}
\usage{
RHS_from_edgemat(caudabox, column)
}
\arguments{
\item{caudabox}{A caudabox object}

\item{column}{PLACEHOLDER}
}
\description{
PLACEHOLDER
}
