% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CauDA.R
\name{get_cols_to_include}
\alias{get_cols_to_include}
\title{PLACEHOLDER}
\usage{
get_cols_to_include(caudabox, column)
}
\arguments{
\item{caudabox}{A caudabox object}

\item{column}{PLACEHOLDER}
}
\description{
PLACEHOLDER
}
